requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4"
azure-core==1.36.0 ; python_version >= "3.11" and python_version < "4"
azure-storage-blob==12.26.0 ; python_version >= "3.11" and python_version < "4"
certifi==2025.10.5 ; python_version >= "3.11" and python_version < "4"
cffi==2.0.0 ; python_version >= "3.11" and python_version < "4" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.4 ; python_version >= "3.11" and python_version < "4"
cryptography==46.0.2 ; python_version >= "3.11" and python_version < "4"
idna==3.11 ; python_version >= "3.11" and python_version < "4"
isodate==0.7.2 ; python_version >= "3.11" and python_version < "4"
multidict==6.7.0 ; python_version >= "3.11" and python_version < "4"
propcache==0.4.1 ; python_version >= "3.11" and python_version < "4"
pycparser==2.23 ; python_version >= "3.11" and python_version < "4" and platform_python_implementation != "PyPy" and implementation_name != "PyPy"
pydantic-core==2.41.4 ; python_version >= "3.11" and python_version < "4"
pydantic-settings==2.11.0 ; python_version >= "3.11" and python_version < "4"
pydantic==2.12.2 ; python_version >= "3.11" and python_version < "4"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_version < "4"
requests==2.32.5 ; python_version >= "3.11" and python_version < "4"
str-or-none==0.1.0 ; python_version >= "3.11" and python_version < "4"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_version < "4"
urllib3==2.5.0 ; python_version >= "3.11" and python_version < "4"
yarl==1.22.0 ; python_version >= "3.11" and python_version < "4"
